+++ torchcsprng/csrc/block_cipher.h
@@ -127,7 +127,7 @@ template<int block_size, typename cipher_t, typename input_index_calc_t, typenam
 void block_cipher(
     void* input_ptr, int64_t input_numel, int input_type_size, input_index_calc_t input_index_calc,
     void* output_ptr, int64_t output_numel, int output_type_size, output_index_calc_t output_index_calc,
-    Device device, cipher_t cipher, int output_elem_per_block, transform_t transform_func) {
+    at::Device device, cipher_t cipher, int output_elem_per_block, transform_t transform_func) {
   if (output_ptr == nullptr || output_numel == 0) {
     return;
   }
@@ -161,7 +161,7 @@ void block_cipher(
 }
 
 template<int block_size, typename cipher_t>
-void block_cipher(Tensor input, Tensor output, cipher_t cipher) {
+void block_cipher(at::Tensor input, at::Tensor output, cipher_t cipher) {
   const auto input_ptr = input.data_ptr();
   const auto input_numel = input.numel();
 
@@ -171,7 +171,7 @@ void block_cipher(Tensor input, Tensor output, cipher_t cipher) {
   }
 
   const auto input_type_size = input.element_size();
-  const auto input_offset_calc = make_offset_calculator<1>(TensorIterator::nullary_op(input));
+  const auto input_offset_calc = make_offset_calculator<1>(at::TensorIterator::nullary_op(input));
   const auto input_index_calc = [input_offset_calc] TORCH_CSPRNG_HOST_DEVICE (uint32_t li) -> uint32_t {
     return input_offset_calc.get(li)[0];
   };
@@ -185,14 +185,14 @@ void block_cipher(Tensor input, Tensor output, cipher_t cipher) {
   }
 
   const auto output_type_size = output.element_size();
-  const auto output_offset_calc = make_offset_calculator<1>(TensorIterator::nullary_op(output));
+  const auto output_offset_calc = make_offset_calculator<1>(at::TensorIterator::nullary_op(output));
   const auto output_index_calc = [output_offset_calc] TORCH_CSPRNG_HOST_DEVICE (uint32_t li) -> uint32_t {
     return output_offset_calc.get(li)[0];
   };
 
   const auto device = output.device();
 
-  block_cipher<block_size>(
+  torch::csprng::block_cipher<block_size>(
       input_ptr, input_numel, input_type_size, input_index_calc,
       output_ptr, output_numel, output_type_size, output_index_calc,
       device, cipher, block_size / output_type_size,
