+++ torchcsprng/csrc/kernels.cpp
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+#include "kernels_header.h"
+#include "kernels_commons.h"
+
+namespace torch {
+namespace csprng {
+namespace cpu {
+
+#include "kernels_body.h"
+
+}}}
